+++ sdk/elastic/azure-mgmt-elastic/azure/mgmt/elastic/aio/operations/_monitored_resources_operations.py
@@ -45,7 +45,7 @@ class MonitoredResourcesOperations:
         self,
         resource_group_name: str,
         monitor_name: str,
-        **kwargs
+        **kwargs: Any
     ) -> AsyncIterable["_models.MonitoredResourceListResponse"]:
         """List the resources currently being monitored by the Elastic monitor resource.
 
@@ -66,7 +66,7 @@ class MonitoredResourcesOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-07-01-preview"
+        api_version = "2020-07-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):