+++ sdk/monitor/azure-monitor-query/samples/sample_batch_query.py
@@ -4,7 +4,7 @@
 from datetime import datetime, timedelta
 import os
 import pandas as pd
-from azure.monitor.query import LogsQueryClient, LogsBatchQueryRequest
+from azure.monitor.query import LogsQueryClient, LogsBatchQuery
 from azure.identity import DefaultAzureCredential
 
 
@@ -12,36 +12,34 @@ credential  = DefaultAzureCredential()
 
 client = LogsQueryClient(credential)
 
-# [START send_batch_query]
+# [START send_query_batch]
 requests = [
-    LogsBatchQueryRequest(
+    LogsBatchQuery(
         query="AzureActivity | summarize count()",
         duration=timedelta(hours=1),
         workspace_id= os.environ['LOG_WORKSPACE_ID']
     ),
-    LogsBatchQueryRequest(
+    LogsBatchQuery(
         query= """AppRequests | take 10  |
             summarize avgRequestDuration=avg(DurationMs) by bin(TimeGenerated, 10m), _ResourceId""",
         duration=timedelta(hours=1),
         start_time=datetime(2021, 6, 2),
         workspace_id= os.environ['LOG_WORKSPACE_ID']
     ),
-    LogsBatchQueryRequest(
-        query= "AppRequests | take 5",
+    LogsBatchQuery(
+        query= "AppRequestss | take 5",
         workspace_id= os.environ['LOG_WORKSPACE_ID'],
         include_statistics=True
     ),
 ]
-responses = client.batch_query(requests)
+responses = client.query_batch(requests)
 
 for response in responses:
-    body = response.body
-    print(response.id)
-    if not body.tables:
-        print("Something is wrong")
-    else:
-        for table in body.tables:
-            df = pd.DataFrame(table.rows, columns=[col.name for col in table.columns])
-            print(df)
+    try:
+        table = response.tables[0]
+        df = pd.DataFrame(table.rows, columns=[col.name for col in table.columns])
+        print(df)
+    except TypeError:
+        print(response.error)
 
-# [END send_batch_query]
+# [END send_query_batch]