+++ sdk/monitor/azure-monitor-query/tests/async/test_logs_client_async.py
@@ -2,7 +2,7 @@ import pytest
 import os
 from azure.identity.aio import ClientSecretCredential
 from azure.core.exceptions import HttpResponseError
-from azure.monitor.query import LogsBatchQueryRequest
+from azure.monitor.query import LogsBatchQuery
 from azure.monitor.query.aio import LogsQueryClient
 
 def _credential():
@@ -40,27 +40,27 @@ async def test_logs_server_timeout():
         assert e.message.contains('Gateway timeout')
 
 @pytest.mark.live_test_only
-async def test_logs_batch_query():
+async def test_logs_query_batch():
     client = LogsQueryClient(_credential())
 
     requests = [
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query="AzureActivity | summarize count()",
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query= """AppRequests | take 10  |
                 summarize avgRequestDuration=avg(DurationMs) by bin(TimeGenerated, 10m), _ResourceId""",
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query= "AppRequests | take 2",
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
     ]
-    response = await client.batch_query(requests)
+    response = await client.query_batch(requests)
 
     assert len(response) == 3
 
@@ -85,32 +85,32 @@ async def test_logs_single_query_additional_workspaces_async():
 @pytest.mark.skip('https://github.com/Azure/azure-sdk-for-python/issues/19382')
 @pytest.mark.live_test_only
 @pytest.mark.asyncio
-async def test_logs_batch_query_additional_workspaces():
+async def test_logs_query_batch_additional_workspaces():
     client = LogsQueryClient(_credential())
     query = "union * | where TimeGenerated > ago(100d) | project TenantId | summarize count() by TenantId"
 
     requests = [
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query,
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]
         ),
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query,
             timespan="PT1H",
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]
         ),
-        LogsBatchQueryRequest(
+        LogsBatchQuery(
             query,
             workspace_id= os.environ['LOG_WORKSPACE_ID'],
             additional_workspaces=[os.environ['SECONDARY_WORKSPACE_ID']]
         ),
     ]
-    response = await client.batch_query(requests)
+    response = await client.query_batch(requests)
 
     assert len(response) == 3
 
     for resp in response:
-        assert len(resp.body.tables[0].rows) == 2
+        assert len(resp.tables[0].rows) == 2